# docs/mvp/API_SPECIFICATION_MVP.yaml
...
openapi: 3.0.3
info:
  title: Courtesy Inspection API - MVP
  description: |
    # Courtesy Inspection API Documentation - MVP Version
    
    Essential API specification for the Courtesy Inspection digital vehicle inspection platform.
    This MVP version focuses on core CRUD operations and basic workflow functionality.
    
    ## Authentication
    All API requests require authentication using JWT tokens obtained from the `/auth/login` endpoint.
    Include the token in the `Authorization` header: `Bearer <token>`
    
    ## Rate Limiting
    - Anonymous: 100 requests/hour
    - Authenticated: 1,000 requests/hour
    
    ## Pagination
    List endpoints support cursor-based pagination with `limit` and `cursor` parameters.
    
    ## Error Handling
    Errors follow RFC 7807 Problem Details format.
  version: 1.0.0-mvp
  contact:
    name: API Support
    email: api@courtesyinspection.com
servers:
  - url: https://api.courtesyinspection.com/v1
    description: Production server
  - url: https://staging-api.courtesyinspection.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Development server
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user and receive JWT tokens
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: mechanic@shop.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: SecurePassword123!
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT access token
                  refresh_token:
                    type: string
                    description: JWT refresh token
                  expires_in:
                    type: integer
                    description: Token expiration time in seconds
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Exchange refresh token for new access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
                  description: Valid refresh token
      responses:
        '200':
          description: New access token generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  expires_in:
                    type: integer
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Invalidate user tokens
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Successfully logged out
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # Inspection Endpoints
  /inspections:
    get:
      tags:
        - Inspections
      summary: List inspections
      description: Get paginated list of inspections with basic filters
      operationId: listInspections
      security:
        - bearerAuth: []
      parameters:
        - name: shop_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by shop ID
        - name: vehicle_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by vehicle ID
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, in_progress, completed, sent]
          description: Filter by inspection status
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/cursor'
      responses:
        '200':
          description: List of inspections
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Inspection'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Inspections
      summary: Create inspection
      description: Start a new vehicle inspection
      operationId: createInspection
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - vehicle_id
                - shop_id
              properties:
                vehicle_id:
                  type: string
                  format: uuid
                shop_id:
                  type: string
                  format: uuid
                customer_concerns:
                  type: string
                  description: Customer reported issues
      responses:
        '201':
          description: Inspection created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inspection'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /inspections/{inspection_id}:
    get:
      tags:
        - Inspections
      summary: Get inspection details
      description: Retrieve complete inspection information
      operationId: getInspection
      security:
        - bearerAuth: []
      parameters:
        - name: inspection_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Inspection details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InspectionDetail'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags:
        - Inspections
      summary: Update inspection
      description: Update inspection information
      operationId: updateInspection
      security:
        - bearerAuth: []
      parameters:
        - name: inspection_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectionUpdate'
      responses:
        '200':
          description: Inspection updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inspection'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /sms/send:
    post:
      tags:
        - SMS
      summary: Send SMS message
      description: Send SMS to customer (Shop Manager only)
      operationId: sendSMS
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - to
                - message
              properties:
                to:
                  type: string
                  description: Customer phone number
                  pattern: '^\+1[0-9]{10}$'
                message:
                  type: string
                  maxLength: 200
                  description: Message content (ci.link/xyz saves 66% on SMS costs)
                inspection_id:
                  type: string
                  format: uuid
                  description: Optional inspection association
      responses:
        '200':
          description: SMS sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message_id:
                    type: string
                    description: Message ID
                  status:
                    type: string
                    enum: [queued, sending, sent, failed]
                  sent_at:
                    type: string
                    format: date-time
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    limit:
      name: limit
      in: query
      description: Number of items to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    cursor:
      name: cursor
      in: query
      description: Cursor for pagination
      schema:
        type: string
  responses:
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: Problem type identifier
        title:
          type: string
          description: Human-readable error title
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Human-readable error details
    Pagination:
      type: object
      properties:
        next_cursor:
          type: string
          nullable: true
        has_more:
          type: boolean
        total:
          type: integer
          description: Total count if available
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        role:
          type: string
          enum: [super_admin, shop_manager, mechanic]
        shop_id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Vehicle:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        vin:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        mileage:
          type: integer
        license_plate:
          type: string
        created_at:
          type: string
          format: date-time
...
    Inspection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        vehicle_id:
          type: string
          format: uuid
        shop_id:
          type: string
          format: uuid
        technician_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [draft, in_progress, completed, sent]
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        sent_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    InspectionDetail:
      allOf:
        - $ref: '#/components/schemas/Inspection'
        - type: object
          properties:
            vehicle:
              $ref: '#/components/schemas/Vehicle'
            technician:
              $ref: '#/components/schemas/User'
            findings:
              type: array
              items:
                $ref: '#/components/schemas/Finding'
    InspectionUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [draft, in_progress, completed]
        customer_concerns:
          type: string
        notes:
          type: string
    Finding:
      type: object
      properties:
        id:
          type: string
          format: uuid
        inspection_id:
          type: string
          format: uuid
        category:
          type: string
        component:
          type: string
        severity:
          type: string
          enum: [green, yellow, red]
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
security:
  - bearerAuth: []